+++ Local Pods/GitHubAPI/GitHubAPI/V3MilestoneResponse.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 public struct V3MilestoneResponse: V3Request {
-    public typealias ResponseType = V3Response<[V3Milestone]>
+    public typealias ResponseType = V3DataResponse<[V3Milestone]>
     public var pathComponents: [String] {
         return ["repos", owner, repo, "milestones"]
     }
